+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/periodic/PeriodicProcessService.scala
@@ -505,10 +505,7 @@ class PeriodicProcessService(
         )
       }
       inputConfigDuringExecutionJsonOpt <- periodicProcessesRepository
-        .fetchInputConfigDuringExecutionJson(
-          processName,
-          versionId,
-        )
+        .fetchInputConfigDuringExecutionJson(deployment.periodicProcess.id)
         .run
       inputConfigDuringExecutionJson = inputConfigDuringExecutionJsonOpt.getOrElse {
         throw new PeriodicProcessException(