+++ utils/schemed-kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/schemedkafka/schemaregistry/universal/UniversalSchemaRegistryClientFactory.scala
@@ -3,7 +3,11 @@ package pl.touk.nussknacker.engine.schemedkafka.schemaregistry.universal
 import pl.touk.nussknacker.engine.kafka.{KafkaUtils, SchemaRegistryClientKafkaConfig}
 import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.azure.AzureSchemaRegistryClientFactory
 import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.confluent.client.CachedConfluentSchemaRegistryClientFactory
-import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.{SchemaRegistryClient, SchemaRegistryClientFactory}
+import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.{
+  EmptySchemaRegistry,
+  SchemaRegistryClient,
+  SchemaRegistryClientFactory
+}
 
 object UniversalSchemaRegistryClientFactory extends UniversalSchemaRegistryClientFactory
 
@@ -12,10 +16,14 @@ class UniversalSchemaRegistryClientFactory extends SchemaRegistryClientFactory {
   override type SchemaRegistryClientT = SchemaRegistryClient
 
   override def create(config: SchemaRegistryClientKafkaConfig): SchemaRegistryClientT = {
-    if (config.kafkaProperties.get("schema.registry.url").exists(_.endsWith(KafkaUtils.azureEventHubsUrl))) {
-      AzureSchemaRegistryClientFactory.create(config)
-    } else {
-      CachedConfluentSchemaRegistryClientFactory.create(config)
+    config.kafkaProperties.get("schema.registry.url") match {
+      case None => EmptySchemaRegistry
+      case Some(url) =>
+        if (url.endsWith(KafkaUtils.azureEventHubsUrl)) {
+          AzureSchemaRegistryClientFactory.create(config)
+        } else {
+          CachedConfluentSchemaRegistryClientFactory.create(config)
+        }
     }
   }